chat-with-db-agents/requirements.txt
llama-index-core
llama-index-llms-openai
openai
requests
python-dotenv
streamlit
streamlit_chat
langchain
langchain_openai
langgraph
crewai
